Create a new keyring and create a new account.
The key ring is now created with a static hash (KeyRingID) and
a random key ring identifier. If a new account is created, Uh
is now calculated right (finally) using the key ring ID. This
means that Uh is different for each key ring.
The keys are stored as real hex numbers in the key ring.
There are much changes in the key ring code as to key calculations
and storage. Anyway, a new key can be stored and updated now.
Extra listners are created for allowing more keys to be tried,
before giving up.
Data that lived in the 'lib' subdirectory has been moved back to
the 'data' subdirectory, with a test for the existence of the
'exports' variable.
The web pages are trimmed a bit and should now require no scripting
whatsoever. This has not been fully tested, but I am almost sure.
There could be some optimization between the output of the HSM
shell script and the use of the values (given as string without
'0x') or as a real value (with '0x').